+++ crates/mimic/src/db/types/data.rs
@@ -1,7 +1,8 @@
 use crate::db::types::SortKey;
-use candid::CandidType;
-use icu::impl_storable_unbounded;
+use candid::{CandidType, Decode, Encode};
+use icu::ic::structures::{Storable, storable::Bound};
 use serde::{Deserialize, Serialize};
+use std::borrow::Cow;
 
 ///
 /// DataRow
@@ -32,7 +33,58 @@ pub struct DataValue {
     pub metadata: Metadata,
 }
 
-impl_storable_unbounded!(DataValue);
+impl Storable for DataValue {
+    const BOUND: Bound = Bound::Unbounded;
+
+    fn to_bytes(&self) -> Cow<'_, [u8]> {
+        let mut out = Vec::new();
+
+        // Write blob
+        let blob_bytes = self.bytes.to_bytes();
+        write_chunk(&mut out, &blob_bytes);
+
+        // Write path
+        write_chunk(&mut out, self.path.as_bytes());
+
+        // Write metadata
+        let meta_bytes = Encode!(&self.metadata).expect("encode metadata");
+        write_chunk(&mut out, &meta_bytes);
+
+        Cow::Owned(out)
+    }
+
+    fn from_bytes(bytes: Cow<[u8]>) -> Self {
+        let mut cursor = &bytes[..];
+
+        let bytes = read_chunk(&mut cursor);
+        let path = String::from_utf8(read_chunk(&mut cursor)).expect("invalid utf-8 path");
+        let metadata_buf = read_chunk(&mut cursor);
+        let metadata = Decode!(&metadata_buf, Metadata).expect("decode metadata");
+
+        DataValue {
+            bytes,
+            path,
+            metadata,
+        }
+    }
+}
+
+// read_chunk
+fn read_chunk(buf: &mut &[u8]) -> Vec<u8> {
+    let len = u32::from_le_bytes(buf[..4].try_into().unwrap()) as usize;
+    let val = buf[4..4 + len].to_vec();
+    *buf = &buf[4 + len..];
+
+    val
+}
+
+// write_chunk
+fn write_chunk(buf: &mut Vec<u8>, data: &[u8]) {
+    let len = data.len() as u32;
+    buf.extend(&len.to_le_bytes());
+
+    buf.extend(data);
+}
 
 ///
 /// Metadata